+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Support to convert Kerlink Wanesy CSV format to The Things Stack JSON.
+
 ### Changed
 
 ### Deprecated